Style homepage and partially the quesadilla page
Began styling the homepage, adjusting classes and
style markers. All class selectors, elements, and
attributes currently used correctly. Needs more
styling on quesadilla page. Possibly move the
picture to the left and all of the recipe story to
the left, as if in columns, and then continue with
directions and ingredients underneath that.